Silence compiler warnings about signedness comparison issues. (#1009)
warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

+++ collector/lib/ProfilerHandler.cpp
@@ -44,7 +44,7 @@ bool ProfilerHandler::SendHeapProfile(struct mg_connection* conn) {
     if (mg_send_http_ok(conn, "application/octet-stream", heap_profile_length_) < 0) {
       return false;
     }
-    if (mg_write(conn, heap_profile_.get(), heap_profile_length_) != heap_profile_length_) {
+    if (mg_write(conn, heap_profile_.get(), heap_profile_length_) != (int)heap_profile_length_) {
       return false;
     }
   }
